+++ backend/src/main/java/ddangkong/facade/balance/option/dto/BalanceOptionResponse.java
@@ -1,0 +1,13 @@
+package ddangkong.facade.balance.option.dto;
+
+import ddangkong.domain.balance.option.BalanceOption;
+
+public record BalanceOptionResponse(
+        Long optionId,
+        String name
+) {
+
+    public BalanceOptionResponse(BalanceOption balanceOption) {
+        this(balanceOption.getId(), balanceOption.getName());
+    }
+}
