+++ backend/src/main/java/ddangkong/facade/room/balance/roomvote/dto/RoomBalanceVoteResponse.java
@@ -1,0 +1,12 @@
+package ddangkong.facade.room.balance.roomvote.dto;
+
+import ddangkong.domain.room.balance.roomvote.RoomBalanceVote;
+
+public record RoomBalanceVoteResponse(
+        Long optionId
+) {
+
+    public RoomBalanceVoteResponse(RoomBalanceVote roomBalanceVote) {
+        this(roomBalanceVote.getOptionId());
+    }
+}
