+++ backend/src/main/java/ddangkong/facade/room/balance/roomvote/dto/RoomBalanceVoteResultResponse.java
@@ -1,0 +1,9 @@
+package ddangkong.facade.room.balance.roomvote.dto;
+
+import ddangkong.facade.balance.vote.dto.ContentTotalBalanceVoteResponse;
+
+public record RoomBalanceVoteResultResponse(
+        ContentRoomBalanceVoteResponse group,
+        ContentTotalBalanceVoteResponse total
+) {
+}
