+++ backend/src/main/java/ddangkong/facade/room/dto/RoomJoinRequest.java
@@ -1,0 +1,9 @@
+package ddangkong.facade.room.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RoomJoinRequest(
+        @NotBlank
+        String nickname
+) { // todo validation 닉네임 정책
+}
