+++ backend/src/main/java/ddangkong/facade/room/dto/RoomJoinResponse.java
@@ -1,0 +1,9 @@
+package ddangkong.facade.room.dto;
+
+import ddangkong.facade.room.member.dto.MemberResponse;
+
+public record RoomJoinResponse(
+        Long roomId,
+        MemberResponse member
+) {
+}
