+++ backend/src/main/java/ddangkong/facade/room/dto/RoomSettingRequest.java
@@ -1,0 +1,10 @@
+package ddangkong.facade.room.dto;
+
+import ddangkong.domain.balance.content.Category;
+
+public record RoomSettingRequest(
+        int totalRound,
+        int timeLimit,
+        Category category
+) {  // todo validation
+}
