+++ backend/src/main/java/ddangkong/facade/room/dto/RoomSettingResponse.java
@@ -1,0 +1,15 @@
+package ddangkong.facade.room.dto;
+
+import ddangkong.domain.balance.content.Category;
+import ddangkong.domain.room.Room;
+
+public record RoomSettingResponse(
+        int totalRound,
+        int timeLimit,
+        Category category
+) {
+
+    public RoomSettingResponse(Room room) {
+        this(room.getTotalRound(), room.getTimeLimit(), room.getCategory());
+    }
+}
