+++ backend/src/main/java/ddangkong/facade/room/dto/RoundFinishedResponse.java
@@ -1,0 +1,7 @@
+package ddangkong.facade.room.dto;
+
+public record RoundFinishedResponse(
+        boolean isRoundFinished,
+        boolean isGameFinished
+) {
+}
