+++ backend/src/main/java/ddangkong/facade/room/member/dto/MemberResponse.java
@@ -1,0 +1,13 @@
+package ddangkong.facade.room.member.dto;
+
+import ddangkong.domain.room.member.Member;
+
+public record MemberResponse(
+        Long memberId,
+        String nickname,
+        boolean isMaster
+) {
+    public MemberResponse(Member member) {
+        this(member.getId(), member.getNickname(), member.isMaster());
+    }
+}
