--- backend/src/main/java/ddangkong/service/balance/vote/dto/OptionTotalBalanceVoteResponse.java
@@ -1,17 +1,0 @@
-package ddangkong.service.balance.vote.dto;
-
-import ddangkong.domain.balance.option.BalanceOption;
-import ddangkong.util.PercentageCalculator;
-
-public record OptionTotalBalanceVoteResponse(
-        Long optionId,
-        String name,
-        int percent
-) {
-
-    public OptionTotalBalanceVoteResponse(BalanceOption balanceOption, long contentVoteCount, long optionVoteCount) {
-        this(balanceOption.getId(),
-                balanceOption.getName(),
-                PercentageCalculator.calculatePercent(optionVoteCount, contentVoteCount));
-    }
-}
