--- backend/src/main/java/ddangkong/service/room/balance/roomvote/dto/RoomBalanceVoteResultResponse.java
@@ -1,9 +1,0 @@
-package ddangkong.service.room.balance.roomvote.dto;
-
-import ddangkong.service.balance.vote.dto.ContentTotalBalanceVoteResponse;
-
-public record RoomBalanceVoteResultResponse(
-        ContentRoomBalanceVoteResponse group,
-        ContentTotalBalanceVoteResponse total
-) {
-}
