--- backend/src/main/java/ddangkong/service/room/balance/roomvote/dto/VoteFinishedResponse.java
@@ -1,14 +1,0 @@
-package ddangkong.service.room.balance.roomvote.dto;
-
-public record VoteFinishedResponse(
-        boolean isFinished
-) {
-
-    public static VoteFinishedResponse roundFinished() {
-        return new VoteFinishedResponse(true);
-    }
-
-    public static VoteFinishedResponse allVoteFinished(boolean isAllVoteFinished) {
-        return new VoteFinishedResponse(isAllVoteFinished);
-    }
-}
