--- backend/src/main/java/ddangkong/service/room/dto/RoomJoinRequest.java
@@ -1,9 +1,0 @@
-package ddangkong.service.room.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RoomJoinRequest(
-        @NotBlank
-        String nickname
-) { // todo validation 닉네임 정책
-}
