--- backend/src/main/java/ddangkong/service/room/dto/RoomSettingRequest.java
@@ -1,10 +1,0 @@
-package ddangkong.service.room.dto;
-
-import ddangkong.domain.balance.content.Category;
-
-public record RoomSettingRequest(
-        int totalRound,
-        int timeLimit,
-        Category category
-) {  // todo validation
-}
