--- backend/src/main/java/ddangkong/service/room/dto/RoomSettingResponse.java
@@ -1,15 +1,0 @@
-package ddangkong.service.room.dto;
-
-import ddangkong.domain.balance.content.Category;
-import ddangkong.domain.room.Room;
-
-public record RoomSettingResponse(
-        int totalRound,
-        int timeLimit,
-        Category category
-) {
-
-    public RoomSettingResponse(Room room) {
-        this(room.getTotalRound(), room.getTimeLimit(), room.getCategory());
-    }
-}
